pyproject.toml
[project]
name = "investment-strategy"
version = "0.1.0"
description = "Multi-agent paper-trading lab for medium-long-term portfolio research (Spain, Lightyear)."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "Proprietary — personal use" }
authors = [{ name = "Lluis Goris" }]

# ---------------------------------------------------------------------------
# MVP runtime dependencies.
# Versions are loose (compatible-release) so `uv lock` resolves consistent
# pins into uv.lock — the lock file is the source of truth for reproducibility.
# Adding a dep: prefer well-maintained, pure-Python where possible. Anything
# requiring a C/Rust toolchain on Windows must be tested before committing.
# ---------------------------------------------------------------------------
dependencies = [
    # Data validation & schemas (Financial CoT envelopes, agent I/O contracts)
    "pydantic>=2.7,<3",
    "pydantic-settings>=2.3,<3",

    # Numerics & dataframes
    "numpy>=1.26,<3",
    "pandas>=2.2,<3",
    "scipy>=1.13,<2",

    # Analytics cache (regenerable, queryable). DuckDB reads JSONL natively.
    "duckdb>=1.0,<2",

    # Market data sources (free / freemium)
    "yfinance>=0.2.40",          # equities, ETFs prices & fundamentals
    "fredapi>=0.5",              # macro series (yields, spreads, VIX proxies)
    "requests>=2.32,<3",
    "httpx>=0.27,<1",            # async / SEC EDGAR with custom UA
    "feedparser>=6.0,<7",        # RSS parsing for news_scanner (Yahoo+Google)
    "beautifulsoup4>=4.12,<5",   # HTML cleanup for news snippets when needed

    # Multi-agent orchestration (Bull/Bear debate, Reflection loop)
    "langgraph>=1.0,<2",
    "langchain-core>=1.0,<2",
    "langchain-anthropic>=1.0,<2",

    # Alpaca paper trading (Claude Autonomous experiment)
    "alpaca-py>=0.30,<1",

    # Quant & risk
    "scikit-learn>=1.5,<2",
    "statsmodels>=0.14,<1",
    "hmmlearn>=0.3,<1",          # macro regime detection (HMM)
    "PyPortfolioOpt>=1.5,<2",    # HRP, mean-variance, risk parity helpers
    "arch>=7.0,<8",              # GARCH / vol modeling

    # Config & secrets
    "python-dotenv>=1.0,<2",

    # CLI & UX (operator-facing slash commands wrap these)
    "typer>=0.12,<1",
    "rich>=13.7,<14",

    # Reporting (HTML/markdown summaries for the operator)
    "jinja2>=3.1,<4",

    # JSONL utilities (append-only event log)
    "orjson>=3.10,<4",           # fast, deterministic JSON serialization
]

[project.optional-dependencies]
# Heavier / optional integrations — not required for MVP loop to work.
extra = [
    "openbb>=4.3",               # broad market data aggregator (optional)
    "matplotlib>=3.9,<4",        # ad-hoc plots in notebooks
    "seaborn>=0.13,<1",
]
# Technical analyst: pandas-ta is convenient but currently requires
# Python >= 3.12 (we're on 3.11) AND has known pandas 2.x compat
# issues on Windows. Skipping it on purpose — `scripts/technical_analyst.py`
# implements the same indicators (MA, RSI, MACD, Bollinger) directly on
# numpy/pandas with no dependency on pandas-ta.

# HRP optimizer (López de Prado). Optional — `scripts/portfolios/hrp_portfolio.py`
# falls back to inverse-volatility weighting when riskfolio is missing.
hrp = [
    "riskfolio-lib>=5.0",
]

dev = [
    "pytest>=8.2,<9",
    "pytest-cov>=5.0,<6",
    "ruff>=0.5,<1",
    "mypy>=1.10,<2",
    "ipython>=8.26,<9",
    "jupyterlab>=4.2,<5",
]

# ---------------------------------------------------------------------------
# Build / packaging
# Hatchling is the default build backend for `uv init`; src/ layout keeps
# package code separate from data/ and notebooks/, which avoids accidental
# packaging of secrets or large files.
# ---------------------------------------------------------------------------
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ---------------------------------------------------------------------------
# Linting & formatting
# Ruff replaces flake8 + isort + most of pylint. Strict enough to catch real
# bugs (E, F, B, S, RUF) but not pedantic (D for docstrings is OFF — comments
# are dense by convention but not enforced as docstrings).
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",        # pyflakes
    "I",        # isort
    "B",        # bugbear
    "S",        # flake8-bandit (security)
    "UP",       # pyupgrade
    "RUF",      # ruff-specific
]
ignore = [
    "S101",     # asserts allowed in tests
    "E501",     # line length handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S", "B"]

# ---------------------------------------------------------------------------
# Type checking
# Strict on src/, lenient on tests/. Third-party stubs missing for yfinance,
# hmmlearn, PyPortfolioOpt — silence those rather than fight upstream.
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
files = ["src"]

[[tool.mypy.overrides]]
module = ["yfinance.*", "hmmlearn.*", "pypfopt.*", "fredapi.*", "arch.*"]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "slow: tests that hit network or run > 1s",
    "integration: end-to-end paper-portfolio simulations",
]

[dependency-groups]
dev = [
    "anthropic>=0.101.0",
    "plotly>=6.7.0",
    "streamlit>=1.57.0",
    "pytest>=8.2,<10",
    "pytest-cov>=5,<10",
]
